Add links in README and a start to git guide
The Git guide is extraodinarly rough and not finished. I have NOT gone through spelling or
Gramar, and I have not finished the majority of the content. I added some text and links
to the guides in README. I also made the "don't panic" in the linux guide big, yellow, and
Friendly.