Started Sphinx Build System
Sphinx should now generate documentation with autodoc for the documentation in
the acmwebsite directory. However, it is very bare-bones, so you need to run
`sphinx-apidoc` and generate the documentation for the `acmwebsite` directory
into a folder in the `docs` folder. In the future, we should create a builder
for the entire documentation.

Documentation will be imperative for new people to use this library. This and
more documentation has been started on the branch `docs` and partly
`profile-page`.

In the `setup.py`, the optional dependency `sphinx` was added under
`documentation`. In addition, `.gitignore` has added the Sphinx build directories.

Resolves: #58.