Modify Sphinx Makefile for `api` Folder and Fix Misc. Doc Errors
The Sphinx Makefile required running the `sphinx-apidoc` manually in all cases.
However, now `make html` should compile the documentation sources into a `api`
folder that Sphinx uses to build the *true* documentation.

In building this, a few errors appeared while building the documentation. Specifically, the
`helpers.py` `ftime` documentation used the `:param:` directive in the title,
which is not allowed.

Sphinx Autodoc could not import the module `PIL`, so an `autodoc_mock_imports`
list was added to the `conf.py` in the docs. More warnings have appeared, but
they should be corrected once the Sphinx documentation expands beyond our
docstrings.