Change Makefile Behavior Mildly and Fix Extra in `setup.py`
The Makefile will now always compile the apidocs when called with html. This is
not entirely as desired, so we should make a better build system so that it only
re-calls `sphinx-apidoc` when changes to the code are made. `./api` has also
become `./api/` so it is obvious that it is a directory.

In `setup.py`, using `pip install -e .[doc]` will now install `sphinx` with pip.
It is recommended you use this in a virtual environment. Previously,
extras_require was not working as desired, as well as not adhering to `PEP508`.
This current method *should* work; in some cases (David), however, autodoc
complains and won't compile... yet.